Reorganized YMAP operators and UI
